# server/openapi.yaml
openapi: 3.0.3
info:
  title: ShopPal Core API
  description: >
    Core e-commerce API for ShopPal — authentication, admin authentication,
    users, products, carts, orders and reviews. Session auth uses a
    short-lived JWT access token (returned in the response body) plus an
    HttpOnly refresh-token cookie. Admin sessions use a parallel /admin
    login flow whose access tokens carry a role: admin claim, required by
    the product create/update/delete endpoints.
  version: "1.0.0"
  contact:
    name: ShopPal
...
servers:
  - url: /
    description: Same-origin (behind ingress / reverse proxy)
security:
  - bearerAuth: []
paths:
...
  /product:
    get:
      tags: [Products]
      summary: List products (filterable, sortable, paginated)
      security: []
      parameters:
        - name: limit
          in: query
          schema: { type: integer, minimum: 1, default: 10 }
        - name: offset
          in: query
          schema: { type: integer, minimum: 0, default: 0 }
        - name: sortBy
          in: query
          schema:
            type: string
            enum: [title, price, rating, stock, brand]
            default: title
        - name: order
          in: query
          schema: { type: string, enum: [asc, desc], default: asc }
        - name: categories
          in: query
          description: Comma-separated `Category` values
          schema: { type: string }
        - name: title
          in: query
          schema: { type: string }
        - name: brand
          in: query
          schema: { type: string }
        - name: minPrice
          in: query
          schema: { type: number, minimum: 0, default: 0 }
        - name: maxPrice
          in: query
          schema: { type: number, minimum: 0 }
        - name: minRating
          in: query
          schema: { type: number, minimum: 0, maximum: 5, default: 0 }
        - name: inStock
          in: query
          schema: { type: boolean }
      responses:
        "200":
          description: Product page
          content:
            application/json:
              schema:
                type: object
                properties:
                  products:
                    type: array
                    items: { $ref: "#/components/schemas/Product" }
        "400":
          $ref: "#/components/responses/ValidationError"
    post:
      tags: [Products]
      summary: Create a product
      description: Admin only — used by the `admin` frontend. Requires an admin access token (see the Admin tag).
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/ProductInput" }
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                type: object
                properties:
                  product: { $ref: "#/components/schemas/Product" }
        "400":
          $ref: "#/components/responses/ValidationError"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
...
  /cart:
    post:
      tags: [Carts]
      summary: Resolve guest-cart product IDs to full product records
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [product_ids]
              properties:
                product_ids:
                  type: array
                  items: { type: string, format: uuid }
      responses:
        "200":
          description: Products in the guest cart
          content:
            application/json:
              schema:
                type: object
                properties:
                  products:
                    type: array
                    items: { $ref: "#/components/schemas/Product" }
...
  /order:
    post:
      tags: [Orders]
      summary: Create an order as a guest
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CreateOrderRequest" }
      responses:
        "201":
          description: Order created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Order" }
        "400":
          $ref: "#/components/responses/ValidationError"
    patch:
      tags: [Orders]
      summary: Bulk-update product stock levels (post-checkout reconciliation)
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [product_ids, new_stocks]
              properties:
                product_ids:
                  type: array
                  items: { type: string, format: uuid }
                new_stocks:
                  type: array
                  items: { type: integer, minimum: 0 }
      responses:
        "204":
          description: Stocks updated
  /order/{user_id}:
    get:
      tags: [Orders]
      summary: List the authenticated user's orders
      parameters:
        - $ref: "#/components/parameters/UserId"
        - name: limit
          in: query
          schema: { type: integer, minimum: 1, maximum: 30 }
        - name: offset
          in: query
          schema: { type: integer, minimum: 0 }
      responses:
        "200":
          description: Order history page
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/OrderHistoryItem" }
        "401":
          $ref: "#/components/responses/Unauthorized"
    post:
      tags: [Orders]
      summary: Create an order for the authenticated user
      parameters:
        - $ref: "#/components/parameters/UserId"
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CreateOrderRequest" }
      responses:
        "201":
          description: Order created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Order" }
        "400":
          $ref: "#/components/responses/ValidationError"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /order/{order_id}:
    patch:
      tags: [Orders]
      summary: Update an order's status or delivery address
      parameters:
        - $ref: "#/components/parameters/OrderId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [user_id]
              properties:
                user_id: { type: string, format: uuid }
                order_status: { $ref: "#/components/schemas/OrderStatus" }
                delivery_address: { $ref: "#/components/schemas/Address" }
      responses:
        "204":
          description: Updated
        "401":
          $ref: "#/components/responses/Unauthorized"
    delete:
      tags: [Orders]
      summary: Cancel/delete a recently-placed order
      parameters:
        - $ref: "#/components/parameters/OrderId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [user_id]
              properties:
                user_id: { type: string, format: uuid }
      responses:
        "204":
          description: Deleted
        "401":
          $ref: "#/components/responses/Unauthorized"
...
  /coupon:
    get:
      tags: [Coupons]
      summary: List active coupons currently redeemable
      security: []
      responses: { "200": { description: OK } }
    post:
      tags: [Coupons]
      summary: Create a coupon (admin only)
      responses:
        "201": { description: Created }
        "403": { description: Forbidden }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: >
        Access token returned from signup/login/refresh. Sent as
        `Authorization: Bearer <token>`. A parallel HttpOnly `refreshToken`
        cookie is used by `/auth/refresh-token` to mint new access tokens.
  parameters:
    UserId:
      name: user_id
      in: path
      required: true
      description: Must match the authenticated user's ID.
      schema: { type: string, format: uuid }
...
    OrderId:
      name: order_id
      in: path
      required: true
      schema: { type: string, format: uuid }
  responses:
    ValidationError:
      description: Request failed validation
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorResponse" }
    Unauthorized:
      description: Missing/invalid access token, or acting on another user's resource
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorResponse" }
    Forbidden:
      description: Valid access token, but it doesn't carry the required admin role
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorResponse" }
...
  schemas:
    ErrorResponse:
      type: object
      properties:
        type: { type: string, example: "Bad Request" }
        message: { type: string }
        statusCode: { type: integer, example: 400 }
    Address:
      type: object
      required: [country, city, street]
      properties:
        country: { type: string }
        city: { type: string }
        street: { type: string }
...
    Dimension:
      type: object
      properties:
        width: { type: number, nullable: true }
        height: { type: number, nullable: true }
        depth: { type: number, nullable: true }
        weight: { type: number, nullable: true }
...
    Category:
      type: string
      enum:
        - beauty
        - fragrances
        - furniture
        - groceries
        - home-decoration
        - kitchen-accessories
        - laptops
        - mens-shirts
        - mens-shoes
        - mens-watches
        - mobile-accessories
        - motorcycle
        - skin-care
        - smartphones
        - sports-accessories
        - sunglasses
        - tablets
        - tops
        - vehicle
        - womens-bags
        - womens-dresses
        - womens-jewellery
        - womens-shoes
        - womens-watches
    OrderStatus:
      type: string
      enum: [arrived, purchased, delivering]
...
    Product:
      type: object
      properties:
        product_id: { type: string, format: uuid }
        title: { type: string }
        description: { type: string, nullable: true }
        category: { $ref: "#/components/schemas/Category" }
        price: { type: number }
        rating: { type: number, minimum: 0, maximum: 5 }
        stock: { type: integer }
        images:
          type: array
          items: { type: string, format: uri }
          nullable: true
        thumbnail: { type: string, format: uri, nullable: true }
        brand: { type: string, nullable: true }
        return_policy: { type: string, nullable: true }
        shipping_info: { type: string, nullable: true }
        warranty_info: { type: string, nullable: true }
        dimension: { $ref: "#/components/schemas/Dimension" }
    ProductInput:
      allOf:
        - $ref: "#/components/schemas/Product"
      required: [title, price, rating, stock, category]
...
    CreateOrderRequest:
      type: object
      required: [product_ids, quantities, billing_info, delivery_address]
      properties:
        product_ids:
          type: array
          items: { type: string, format: uuid }
          minItems: 1
        quantities:
          type: array
          items: { type: integer, minimum: 1 }
          minItems: 1
          description: Same length and order as `product_ids`.
        billing_info: { type: string }
        delivery_address: { $ref: "#/components/schemas/Address" }
    Order:
      type: object
      properties:
        order_id: { type: string, format: uuid }
        issued_time: { type: string, format: date-time }
        order_status: { $ref: "#/components/schemas/OrderStatus" }
        billing_info: { type: string, nullable: true }
        delivery_address: { $ref: "#/components/schemas/Address" }
    OrderHistoryItem:
      type: object
      properties:
        order_id: { type: string, format: uuid }
        issued_time: { type: string, format: date-time }
        order_status: { $ref: "#/components/schemas/OrderStatus" }
        delivery_address: { $ref: "#/components/schemas/Address" }
        billing_info: { type: string, nullable: true }
        products:
          type: array
          items:
            type: object
            properties:
              product_id: { type: string, format: uuid }
              thumbnail: { type: string, format: uri, nullable: true }
              title: { type: string }
              price: { type: number }
              quantity: { type: integer }
...