# server/src/app.ts
import express from 'express';
import 'reflect-metadata';
import cookieParser from 'cookie-parser';
import cors from 'cors';

import * as dotenv from 'dotenv';
dotenv.config();

import { AppDataSource } from '@/shared/db/pg.data-source';
import UserRouter from '@/api/users/user.router';
import AuthRouter from '@/api/auth/auth.router';
import AdminRouter from '@/api/admin/admin.router';
import OrderRouter from '@/api/orders/order.router';
import CartRouter from '@/api/carts/carts.router';
import ReviewRouter from '@/api/reviews/review.router';
import ProductRouter from '@/api/products/product.router';
import WishlistRouter from '@/api/wishlist/wishlist.router';
import AddressRouter from '@/api/addresses/addresses.router';
import PaymentMethodRouter from '@/api/payment-methods/payment-methods.router';
import CouponRouter from '@/api/coupons/coupons.router';
import OrderTrackingRouter from '@/api/order-tracking/order-tracking.router';
import QnaRouter from '@/api/qna/qna.router';
import RecentlyViewedRouter from '@/api/recently-viewed/recently-viewed.router';
import NotificationsInboxRouter from '@/api/notifications-inbox/notifications-inbox.router';
import { errorMiddleware } from '@/middlewares/error.middleware';
import corsMiddleware from './middlewares/cors.middleware';
import { mountApiDocs } from './docs';

// Express app setup:
const app = express();
const PORT = 3000;

// Middlewares (before routes):
app.use(
	process.env.NODE_ENV === 'production'
		? cors({
				origin: [process.env.CLIENT_PROD_URL, process.env.ADMIN_PROD_URL].filter(
					(origin): origin is string => !!origin
				),
				credentials: true,
			})
		: corsMiddleware
);
app.use(express.json());
app.use(cookieParser());

// Health & readiness (referenced by container healthcheck + k8s probes):
app.get('/health', (_req, res) => {
	res.status(200).json({ status: 'ok', service: 'server' });
});
app.get('/ready', async (_req, res) => {
	try {
		if (AppDataSource.isInitialized) {
			await AppDataSource.query('SELECT 1');
		}
		res.status(200).json({ status: 'ready' });
	} catch (err: any) {
		res.status(503).json({ status: 'not-ready', error: err?.message });
	}
});

// API docs (OpenAPI spec + Swagger UI):
mountApiDocs(app);

// Routes:
app.use('/user', UserRouter);
app.use('/auth', AuthRouter);
app.use('/admin', AdminRouter);
app.use('/order', OrderRouter);
app.use('/cart', CartRouter);
app.use('/review', ReviewRouter);
app.use('/product', ProductRouter);
app.use('/wishlist', WishlistRouter);
app.use('/address', AddressRouter);
app.use('/payment-method', PaymentMethodRouter);
app.use('/coupon', CouponRouter);
app.use('/tracking', OrderTrackingRouter);
app.use('/qna', QnaRouter);
app.use('/recently-viewed', RecentlyViewedRouter);
app.use('/notifications', NotificationsInboxRouter);

// Middlewares (after routes):
app.use(errorMiddleware); // Every endpoint & middleware chain avaliable will eventually go to error middleware in case of exception thrown

//  !!!!!!!! npm run start:dev !!!!!!!!
app.listen(PORT, () => {
	console.log(`Server is running on port ${PORT}`);

	// Initialize database connection:
	AppDataSource.initialize()
		.then(() => console.log('Database is up and running'))
		.catch((error) => console.log(`error starting database connection -> ${error.message}`));
});